+++ Content.Server/_Funkystation/Tools/Components/CheapLighterComponent.cs
@@ -2,29 +2,31 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-using Robust.Shared.GameStates;
 using Robust.Shared.Audio;
 
-namespace Content.Shared.Tools.Components;
+namespace Content.Server._Funkystation.Tools.Components;
 
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
+[RegisterComponent]
 public sealed partial class CheapLighterComponent : Component
 {
     /// <summary>
     /// An additional sound the lighter should play when switched on, which can be interrupted when it's closed.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField, AutoNetworkedField]
+    [ViewVariables(VVAccess.ReadWrite), DataField]
     public SoundSpecifier? SoundActivate;
 
     /// <summary>
     /// The chance the lighter will fail to light, between 0.0 and 1.0.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField, AutoNetworkedField]
+    [ViewVariables(VVAccess.ReadWrite), DataField]
     public float FailChance = 0f;
 
     /// <summary>
     /// The sound the lighter will play when it fails to light.
     /// </summary>
-    [ViewVariables(VVAccess.ReadWrite), DataField, AutoNetworkedField]
+    [ViewVariables(VVAccess.ReadWrite), DataField]
     public SoundSpecifier? SoundFail;
+
+    [DataField]
+    public EntityUid? LighterStream;
 }
